# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Education System API",
    "version": "1.0.0",
    "description": "Backend API for the Education System project. All responses use the wrapper `{ success, data, error }`. Success: `success: true`, `data: <payload>`, `error: null`. Failure: `success: false`, `data: null`, `error: { code, message, details }`. Every route is served under both /api and /api/v1 for one release. List endpoints share one pagination envelope { items, pagination: { page, limit, total, totalPages } } with limit max 50 and default 10. Common error codes: VALIDATION_ERROR (400), INVALID_JSON (400), UNSUPPORTED_FILE_TYPE (400), FILE_REQUIRED (400), UNAUTHORIZED (401), INVALID_CREDENTIALS (401), INVALID_CURRENT_PASSWORD (401), INVALID_RESET_TOKEN (401), INVALID_VERIFICATION_TOKEN (401), REFRESH_TOKEN_EXPIRED (401), REFRESH_TOKEN_REVOKED (401), INVALID_WEBHOOK_SIGNATURE (401), FORBIDDEN (403), SUBSCRIPTION_REQUIRED (403), NO_ACTIVE_SUBSCRIPTION (403), EMAIL_NOT_VERIFIED (403), TEACHER_NOT_FOUND (404), SUBSCRIPTION_NOT_FOUND (404), NOT_FOUND (404), SECTION_NOT_FOUND (404), ACTIVE_SUBSCRIPTION_EXISTS (409), SUBSCRIPTION_NOT_ACTIVE (409), ACTIVE_SUBSCRIPTIONS_EXIST (409), DUPLICATE_FIELD (409), AMBIGUOUS_USERNAME (409), FILE_TOO_LARGE (413), RATE_LIMIT_EXCEEDED (429). ACTIVE_SUBSCRIPTION_EXISTS means the calling student already holds one; ACTIVE_SUBSCRIPTIONS_EXIST means an admin delete is blocked by subscribers."
  },
  "servers": [
...
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AddContent": {
        "type": "object",
        "required": ["type", "title"],
        "properties": {
          "type": { "type": "string", "enum": ["LECTURE", "LESSON_CONTENT", "HOMEWORK"] },
          "title": { "type": "string", "example": "Algebra basics" },
          "body": { "type": "string", "nullable": true },
          "fileUrl": { "type": "string", "nullable": true, "description": "Same-origin /api/files/<name> path or allowlisted host (FILE_URL_ALLOWED_HOSTS)" },
          "isPublished": { "type": "boolean", "default": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/teacher/dashboard/subscribers": {
      "get": {
        "tags": ["Teacher Dashboard"],
        "summary": "List of subscribers for the logged-in teacher (teacher role)",
        "description": "Omitting status applies the active rule (ACTIVE status with endDate in the future). status filters by exact status. Paginated with page/limit (max 50).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ACTIVE", "EXPIRED", "CANCELLED"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50 } }
        ],
        "responses": { "200": { "description": "Subscribers" } }
      }
    },
    "/teacher/dashboard/income": {
      "get": {
        "tags": ["Teacher Dashboard"],
        "summary": "Income for this month, last 3 months and year (teacher role)",
        "description": "Sums paid subscriptions (ACTIVE plus EXPIRED) by startDate. CANCELLED subscriptions are excluded.",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Income summary" } }
      }
    },
    "/teacher/dashboard/content": {
      "post": {
        "tags": ["Teacher Dashboard"],
        "summary": "Add content as a teacher (teacher role)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddContent" } } }
        },
        "responses": { "201": { "description": "Content created" } }
      }
    },
...
  }
}